Fix support for coordination numbers
This commit fixes the calculation of the control number for
coordination numbers. Previously the calculation was done on the
parsed number where the date part of the number had been normalized to
the actual date (i.e sans the added 60 to it.) The test coordination
numbers provided by skatteverket has a control number that is correct
if you do the calculation with the added 60.